Allow android_manifest to be created implicitly
Summary:
The android_manifest rule predates graph enhancement, so we're requiring
people to declare them explicitly, even for trivial apps.  This ends
now.  If manifest_skeleton is specified instead of manifest, we'll
automatically enhance a manifest merge rule.

Note that this uses an entirely different codepath to find the list of
deps to add to the manifest, but the new one is actually more correct.
For example, it is able to traverse a dep chain that goes through C++.

The old android_manifest rule should probably be deprecated, but we
can't stop using it in our FB apps until we add support for
post-processing the merged manifest (kind of like we allow
post-processing class files).

Test Plan: Added integration test.  Looked at docs.

Reviewed By: asp2insp

fbshipit-source-id: 763ff6b