Add cacheable option to genrule
Summary: This adds the ability for a genrule to specify whether its outputs are cacheable or not. In an ideal world, it always would be, but we have run into a couple of use cases where rules are not. This actually is already the case for a couple of the other genrule subtypes. This property should be considered private, and the expectation should be that it may be removed at any time.

Test Plan: Added tests, updated some other tests

Reviewed By: ttsugriy

fbshipit-source-id: 0f7547b