Introduce the 'output' macro and add support for it in CxxLink
Summary:
Introduce the "output" macro which expands to a named supplementary output of
the rule in which it's used.

Implement support for this in the 'linker_flags' field of Cxx rules. This flag
is not supported in other locations like exported_linker_flags.

The supplementary output must be declared via `linker_extra_outputs` of the
rule using the macro.

The output can be addressed in over rules via `$(location :binary[foo])` where
`foo` is the supplementary output name.

Note that due to the complex expansions of cxx / apple libraries and binaries,
these outputs may not be properly forwarded through wrapper rules...

Documentation changes coming in a future diff.

Test Plan:
Added some tests. The integraiton test serve as an example of how this can be
used. The tests cover shared libraries and binaries, both cxx_ and apple_.
Other rule types call into this code path, but I'm not sure if the feature
works there.

Reviewed By: ttsugriy

fbshipit-source-id: 4f7bec0