Make c++ depfiles use the CoveredByDepfilePredicate
Summary:
c++ depfile support currently doesn't use the predicate for whether a
sourcepath is covered by the depfile. This means that, after building,
the c++ rules need to list all of their possible SourcePath inputs that
aren't covered by the depfile.

This adds a PreprocessDelegate.isCoveredByDepfile() function and uses it
as appropriate.

This simplifies the c++ depfile logic and will allow us to remove
Tool/Arg.getInputs().

Test Plan: CI

Reviewed By: yiding

fbshipit-source-id: 2686089