Move ChromeTraceWriter and ChromeTraceBuckConfig to event.chrome_trace package
Summary:
Move `ChromeTraceWriter` and `ChromeTraceBuckConfig` to
`event.chrome_trace` package, but keep `ChromeTraceEventListener`
in the `event.listener` package.

This change is needed to make possible for `build_slave` implementation
depend on `ChromeTraceBuckConfig` which is needed to upload distbuild
chrome traces.

Unfortunately currently exists revernse dependency `listener` from
`build_slave` which doesn't allow `build_slave` to depend on `listener`.

Probably it would be best to remove `build_slave` dependency from
`listener`, but that seems to be not an easy task.

`ChromeTraceBuildListener` cannot be moved to `event.chrome_trace`,
because it has too many dependencies.

Test Plan: CI

Reviewed By: ttsugriy

fbshipit-source-id: 08bca54