Move Watchman.build() into a separate WatchmanFactory class.
Summary:
There was a lot going on in `Watchman.java` and very little of it seemed to
have to do with an instance of the `Watchman` class. Moving the top-level
static factory method (`public static Watchman build()`) into a new
`WatchmanFactory` class put the new `WatchmanFactory.java` file at 462 lines
while the remaining `Watchman.java` class was left at 106 lines. It turns out
that there are only two non-trivial instance methods on `Watchman`, which was
hard to discern when all of the code was in one file.

This revision was brought to you by IntelliJ.

* I selected "Move..." on the `build()` method.
* I checked all the boxes.
* I let IntelliJ do its thing.

The only thing I had to do myself is add `WatchmanWatcher.java` to the `BUCK`
file.

Test Plan: CI

Reviewed By: ttsugriy

fbshipit-source-id: 94375f5