Don't compute OUTPUT_HASH metadata for large artifacts
Summary:
Very large artifacts take a long time to hash, so we exclude them from
input-based rule key checks via `build.input_rule_key_file_size_limit`.  Despite
not using input-based checks for these rules, we were eagerly computing the
output hash and storing it in the artifact as a way to pre-seed the rulekey
cache for said input-based checks!  So, let's stop doing that.

Since OUTPUT_HASH is part of the artifact, we add the value of
`build.input_rule_key_file_size_limit` to all rulekeys, which ensures that we'll
always have (or not have) OUTPUT_HASH for a given rulekey, which keeps our
outputs deterministic.

Test Plan: CI

Reviewed By: dinhviethoa

fbshipit-source-id: 34c038f