Ensure Swift runtime is embedded in targets that need it
Summary: If an Apple bundle (app or test) includes Swift transitively (i.e., not directly), it will not automatically include the Swift runtime which would cause a crash on launch. In this diff we make sure that `ALWAYS_EMBED_SWIFT_STANDARD_LIBRARIES` is always set if a bundle transitively depends on any Swift code.

Test Plan: New test included in diff + manual test on internal project

Reviewed By: dinhviethoa

fbshipit-source-id: 24f24f8