Add a fixed-length header length indicator to multi-file exo install
Summary:
This will allow the native agent to read the entire header in 2 system
calls without risk of reading past the end of the header, so we won't
need to worry about buffering the extra data that is part of the file
(or the next header).

Test Plan: Exopackage test.

fbshipit-source-id: 8e7b496