Add toolchains with capabilities and migrate toolchains that provide CXX platforms to use this feature
Summary:
`ToolchainWithCapabilities` is a generic interface that can be used to access multiple toolchains that shared the same interface.

In this particular change toolchains that povide CXX platforms are migrated to use `CxxPlatformSupplier` interface. These toolchains then are asked to provide CXX platforms when `CxxPlatformsProvider` is created.

Test Plan: CI

Reviewed By: ttsugriy

fbshipit-source-id: d8d6f3e