Fix target graph hashing parallelism
Summary:
I was using FJPs in a way that was not well suited. In particular, FJPs don't allow blocking calls (e.g. the filesystem reads) during the unit of work.
This switches to a bottom-up traversal with a loading cache which maintains the parallelism, but
allows simpler code and concurrency structure.

Test Plan: Existing tests still pass, target node hash is stable, stuff is still parallel.

Reviewed By: ttsugriy

fbshipit-source-id: a0739d7