Change Watchman to vend out new clients instead of store one.
Summary:
Previously, `com.facebook.buck.io.Watchman` had an `Optional<WatchmanClient>`
field and implemented `AutoCloseable`. The `Watchman.getWatchmanClient()` would
return a reference to its internal `Optional<WatchmanClient>` field and
`Watchman.close()` would invoke `close()` on the client, if present.

Things got a bit messy when it came to managing the lifetimes of these objects.
For example, because `Watchman` implemented `AutoCloseable`, it was created
in a try-with-resources in `Main.runMainWithExitCode()` that spanned 500 lines!
Deep within that code, it would create a `WatchmanWatcher` that stored the
`Watchman`'s `WatcherClient` as a field. This was not a good situation because
the `WatchmanWatcher` could easily outlive the `Watchman`, in which case
it could end up with a stale client that was closed by the `Watchman`.

There was an outstanding `TODO` in the `Watchman` class to move
`WatchmanClient` out of the class and to make `Watchman` an immutable
datatype. Overall, that is what I tried to do, but it was tricky because
certain callsites (most notably, `AbstractBuildFileSpec`) have access to
`Watchman`, but cannot easily be changed to get the additional data they
need to construct a `WatchmanClient` on their own. (The default implementation,
`WatchmanTransportClient`, needs a `Console` and a `Clock`.)

Because of these constraints, I decided to change `Watchman` so that
it would maintain references to `Console` and `Clock` (for which lifetime
management issues are less of a concern) and give it a `createClient()`
method that //always// returns a new `WatchmanClient`. The API is very
clear that the caller is responsible for closing the client.

Because `Watchman` no longer has to manage the lifetime of any
resources like file descriptors, it can be passed around freely. Now the
expectation is that clients will create a `WatchmanClient` when they
need it and leverage `AutoCloseable` to clean it up:

```
try (WatchmanClient watchmanClient = watchman.createClient()) {
  // Use watchmanClient.
}
```

(Because `Watchman` is no longer `AutoCloseable`, we are able to
remove the sizeable try-with-resources block in `runMainWithExitCode()`,
allowing us to unindent 500 lines of code!)

One possible downside to this change is whether we are now creating
far more instances of `WatchmanClient` than we did before. Reading
through the code, it appears the answer is "no." Previous to this change,
`Main.runMainWithExitCode()` would always create a new `Watchman`
object: it was not stored in the `Daemon`. It appears that a `Watchman`'s
client was only being requested in two places: in `AbstractBuildFileSpec`
when Watchman is used to get the list of all build files in the project,
and in `WatchmanWatcher` to find all files that were modified since the
last invocation of Buck. Therefore, we are now creating two instances of
`WatchmanClient` per build instead of one, so this does not seem like it
should be a major concern.

(Note: I renamed `WatchmanTest` to `WatchmanFactoryTest` because it
is really testing `WatchmanFactory`. I should have done this as part of
D6369158 when I moved `Watchman.build()` into a separate
`WatchmanFactory` class.)

To make `WatchmanFactoryTest` continue to work, I had to make some very
specific logic in `WatchmanFactory` configurable. To do this, I added a
public, no-arg constructor to `WatchmanFactory` as well as a second
constructor that takes a lambda that allows for overriding the behavior
that needs to be stubbed out for its unit test. I also had to change some
static methods to instance methods so they could leverage the lambda
that is stored as a field.

Test Plan: CI

fbshipit-source-id: bc0c7fb