Make Watchman query requests in parallel in WatchmanWatcher.
Summary:
We have a use case (with Eden) where we have the unfortunate combination of a
repo with a number of cells and not-instantaneous Watchman queries. (The problem
is that, today, when a commit happens, Watchman invalidates quite a bit of state
when watching an Eden repo, so until we fix this, `hg amend` incurs an unfair
Buck build time penalty because it is doing O(cells) slow Watchman queries
serially.)

This revision parallelizes the Watchman queries. Now `WatchmanWatcher.postEvents()`
spins up its own, temporary `ExecutorService` when it is called. In practice,
`postEvents()` appears to be called once per Buck invocation, and a new
`WatchmanWatcher` is created for each Buck invocation, so there does not appear
to be any value in trying to keep the `ExecutorService` around for future use.

Note that `WatchmanClient` is single-threaded, so we end up creating a separate
one for each cell being processed. If there are far more cells than threads, then this
is particularly wasteful because we should really only create one per thread in that
scenario, though this seems unlikely to be an issue, in practice.

Though now that we are more freely creating `WatchmanClient` objects rather than
sharing one across an entire Buck invocation (particularly due to the changes in
D6371584), it is probably worthwhile to make `WatchmanClient` a bit cheaper, if
possible.

Test Plan:
In addition to CI, performed manual testing to verify that the Watchman queries are
now made in parallel.

I did the following (in an Eden mount):

* `buck-beta build eden/...`
* (made a whitespace change to `eden/fs/service/EdenServiceHandler.cpp`)
* `hg am`
* `buck-beta build eden/...`

Here is a sample build trace of this manual test plan //before// this change:
https://our.intern.facebook.com/intern/buck/builds/d98046a5-19a4-475e-99f4-149ede99b0af/trace/default/

With a screenshot of the build trace (for posterity):
{F115721686}

As you can see, the six `check_watchman` calls are done serially on a single
thread, each one taking ~2.3 seconds. (I thought we only had five cells in
fbsource, but this is how I learned we now have a sixth: `xplat/third-party/bazel-skylib`.)

Here is a sample build trace of this manual test plan //after// this change:
https://our.intern.facebook.com/intern/buck/builds/2e2c3bc8-ae97-4b45-84d1-8e41cdcbd0a8/trace/default/:

Again, with a screenshot of the build trace (for posterity):
{F115721792}

As you can see, there were six `check_watchman` build steps, each on its own thread,
running in parallel. The `check_watchman` step for `fbcode` took ~1.8s and included
`EdenServiceHandler.cpp` in the list of modified files, as expected.

Reviewed By: ttsugriy

fbshipit-source-id: de0679e