Cache the configured JavacSpec
Summary:
JavaBuckConfig exposes a JavacSpec based on the BuckConfig. JavacSpec
will in turn expose a Javac tool for compilation. JavacSpec will
internally cache the tool, but JavaBuckConfig doesn't cache the
JavacSpec.

This means that we actually instantiate many instances of the default
Javac tool. This is particularly bad when that tool is an ExternalJavac
backed by a non-target. In that case, we need to invoke the backing
executable just to determine the version.

This change adds memoization of the JavacSpec in the JavaBuckConfig. In
this way, uses of the default javac will all resolve to the same Tool.

We still may instantiate multiple instances of the same logical Tool if
multiple targets use non-default configuration. That'll be much more
difficult to resolve.

Test Plan: CI, `buck targets --show-rulekey <large-app>` and see regression in action graph creation resolved.

Reviewed By: styurin

fbshipit-source-id: 2fa6d9e