Fix rulekey divergence
Summary:
Currently, we add the class names of every BuildRule/AddsToRuleKey (that
is added to a rulekey) to the rulekey.

The classname of an anonymous class depends on the specific version of
javac that compiled the file. This has a small chance of causing rulekey
divergence (depends on how buck is distributed to users).

The classname of a lambda that implements AddsToRuleKey is effectively
random. This has a high chance of causing rulekey divergence.

This changes the rulekey framework to print an error message when adding
an anonymous/synthetic class to rulekeys and instead will add the enclosing class.

It also removes all production instances of using a lambda. Some instances still
exist of using an anonymous class.

The remaining uses should be removed in the future and the error message
should be changed to a hard error.

Test Plan: CI, `buck targets --show-rulekey folly && buck kill && buck targets --show-rulekey folly`.

Reviewed By: sbalabanov

fbshipit-source-id: 7a23eae