Make command_alias tool add the underlying tool to its rulekey
Summary:
Originally, this was not added to the rulekey in an attempt to make it
affect dependents the same on different platforms. That doesn't actually
work correctly though, it still causes their rulekeys to diverge
(because it has to forward the underlying tool's deps+inputs anyway).

This change adds the tool to the rulekey (we're going to have to come up
with a rather different approach if we want to actually not diverge
rulekeys).

This deletes the tests for getInputs() and getDeps() because those are
about to be completely deleted.

Test Plan: CI

Reviewed By: davidaurelio

fbshipit-source-id: 024cf62