Remove builder from CommandRunnerParams
Summary:
We use builder pattern in `CommandRunnerParams`, but we don't actually need it because all fields are either required or `Optional`. Usage of a builder requires more maintenance and compilation doesn't give good signal about missing parameters when adding a new one (need to run tests or builds to get the exception about missing parameters instead of getting a compilation error). Another disadvantage of having a builder is lots of generated code.

The solution is to disable builder and replace existing usages to use static factory method `of`.

Test Plan: CI

Reviewed By: ttsugriy

fbshipit-source-id: e9c10b6