Make Buck tolerant of the case when Watchman is not present.
Summary:
This was a bug I introduced in D6371584. Previous to that revision, we had the
following code inside a try/catch block that caught `HumanReadableException`:

```
Path watchmanPath = exeFinder.getExecutable(WATCHMAN, env).toAbsolutePath();
```

This was in `Watchman.build()` that had a try/catch block that covered the
entire method even though this `getExecutable()` call was seemingly the only
thing that was expected to throw a `HumanReadableException`.

When I refactored it into `WatchmanFactory.determineWatchmanTransportPath()`
in D6371584, I eliminated the try/catch block altogether and just let the new
method throw. I broke up `Watchman.build()` so that it had two separate
try/catch blocks. I preserved the `catch HumanReadableException` in the second
try/catch but not the first. Unfortunately, the first contained the call to
`determineWatchmanTransportPath()`, so I now believe it was the one that really
needed it.

It turns out that `ExecutableFinder` has a more appropriate API,
`getOptionalExecutable()`, that returns an `Optional<Path>` and does not throw,
so I have changed the code to use that and have removed the reliance on catching
`HumanReadableException` altogether.

Test Plan:
Added a new test, `WatchmanFactoryTest.shouldReturnEmptyWatchmanIfNotOnPath()`.
Verified that it failed with the message we were seeing in the bug report before
this change and that the test succeeded as a result of my change.

Reviewed By: ttsugriy

fbshipit-source-id: 4cfcaef