Fix the behavior sh_binary to work with embedded cells buck-out
Summary:
When embedded cells buck-out is enabled and when sh_binary is in a different cell and had a build result as dependency, the build would fail because it would try to refer to a file in the buck out of the root cell, which is outside of the cell path.

This diff fixes it by making all cells available from the resulting script in buck-out/gen through symlinks.
__default__ is a symlink to the root cell.

Test Plan: added unit test + CI

Reviewed By: ttsugriy

fbshipit-source-id: 545ee8e