Implement a struct API in Python DSL.
Summary:
This API will make transition to Skylark smoother and can
also be convenient on its own, so that instead of having to use
namedtuple, Python DSL clients can just use a `struct` instead.

I haven't documented this API for now, because I'm not sure if we want
to publicly advertise it.

Test Plan: CI

Reviewed By: styurin

fbshipit-source-id: d76fb99