Save mapping of class->dexfile and go directly to dex which contains it
Summary:
We already save the list of classes which we've encountered in our dex files.
This change adds in the mapping of the class to the `DexFile` which contains the definition.
This prevents a linear scan when doing a class lookup and allows us to go directly to the
defining file.

Test Plan: `./scripts/exopackage_test_single.sh`

Reviewed By: dreiss

fbshipit-source-id: cded2c8