haskell_library: generate library GHCi rule with correct name
Summary:
The `baseTarget` is stripped of all type/platform flavors and is only really
meant for the factory-style library rules which generate shared/static libs.
Since the GHCi rule is a user-facing, deployable rule, we need to make sure to
use the exact name that was passed in.

Test Plan: Ran `buck run ...#ghci` on a `haskell_library` rule.

Reviewed By: JonCoens

fbshipit-source-id: 904d803