Shave off nodes which are already present in the remote cache.
Summary:
This diff checks the remote cache for nodes that are already uploaded, and skips them while creating the BuildTargetsQueue.
This means we will not be building extra stuff which we is already present in the cache. This also essentially acts as a top-down
pass of the graph, since we don't traverse dependencies of nodes which are already in the cache.

Test Plan: CI

Reviewed By: ruibm

fbshipit-source-id: 409bc46