JsFile: send arguments as JSON instead of pseudo-cli
Summary: This changeset continues the series of change I've done to communicate with the worker using JSON rather than CLI-looking arguments (that are fragile because they don't do any kind of escaping).

Test Plan: The integration tests verify that the output of the bundles is unchanged. Builds over `fbsource` will confirm the change is compatible with the existing JS worker there.

Reviewed By: davidaurelio

fbshipit-source-id: 6d9bdee