Make Lua resources loaded using the current class loader
Summary:
Lua functionality was moved to a module (which is a stand-alone jar) with all its resources, but these resources are accessed using current thread's classloader, not the current class's classloader.

The fix is to use the current class as the loading point and change the location of resources to be relative.

Test Plan: Local testing

Reviewed By: ttsugriy

fbshipit-source-id: 02f1d3d