Fix modulars swift project umbrella header generation issue
Summary:
In a previous refactor, accidentally we check the wrong variable to determine if the user supplied an umbrella header. The result is if the user sets `apple.generate_missing_umbrella_headers`, buck project would overwrite the user's symlinked header. This fixes it and adds a unittest.
Closes https://github.com/facebook/buck/pull/1656

Reviewed By: milend

Pulled By: milend

fbshipit-source-id: ead6f04