Add Android facet to exopackage module
Summary:
Adds the android facet to the exopackage module which contains android code, enabling use of several IJ tools.
This will not affect most buck developers because this module is disabled by default.

Test Plan: Open project in IJ, import the exopackage module, see that the android tools are enabled.

Reviewed By: dreiss

fbshipit-source-id: d8b5ee2