Avoid recreating PluginManager between buck runs
Summary:
`PluginManager` was re-created in `Main` on every buck run, but the actual `PluginClassLoaders` were the same and accessing toolchains failed because of `ClassCastException` (descriptions were loaded using new plugins and toolchains were loaded using the old plugins hence the difference in class loaders).

The fix is to keep `PluginManager` in static variable and avoid re-creating it when it's already present.

Test Plan:
Using buck build by buck:

```
$ buck kill
$ buck clean
$ buck run some_lua_target
```

The last command should not fail.

Reviewed By: ttsugriy

fbshipit-source-id: 91e45b0