Normalize asset paths during MergeAssets
Summary:
Android's `MergeAssets` step tries to find a relative path from a `ProjectFilesystem`.
With `embedded_cells` config, the wrong filesystem is returned for locating an output directory,
leading to an asset path with a bunch of `../../..` segments when it gets relativized. This diff
corrects the symptom by calling `normalize()` on the path before it is added to the asset mapping,
since the only thing we're trying to get out is a relative pathname. I think the team addressing cell
embedding needs to re-think how finding outputs via `ProjectFilesystem` instances works however,
for a more concrete solution.

Test Plan: New integration tests

Reviewed By: dinhviethoa

fbshipit-source-id: f9363bd