Make createNullConsole() return a singleton. Use it in PerBuildState.
Summary:
`buck query set(someset)` currently creates one PerBuildState
per target in someset. (Why it does this is also somewhat suspect and
will be investigated next, but it seemed like a good idea to make
PerBuildState cheap regardless.) When there are lots of targets in
someset, the overhead of creating PerBuildState becomes
significant. It turned out that PerBuildState was using a clowny way
of creating a null console that involved making a lot of PrintStreams,
initializing PrintStreams is relatively costly, and furthermore there
doesn't seem to be any reason that the null console can't be a
singleton.

Test Plan:
Profile evaluation of a buck query with a large set. No
longer significant overhead from creating PerBuildState.

Reviewed By: dinhviethoa, ttsugriy

fbshipit-source-id: a3c9c03