remove avg download speed from the cache
Summary:
The average download speed from the cache is a misleading number and
it is often a cause of confusion from client who think their network is
misbehaving. The goal here is to first remove it and then introduce a reliable
and clear way how we measure this.

Test Plan:
- CI
- manual run:

  Parsing buck files: finished in 1.3 sec (100%)
  Creating action graph: finished in 0.5 sec (100%)
  Downloading... 1701 artifacts, 13.29 Mbytes
  Building... 35.2 sec (30%) 406/1350 jobs, 406 updated, 28.3% cache miss
   - //src-gen:thrift#source-abi... 1.0 sec (running source_abi[0.9 sec])
   - //third-party/java/dx:dx#source-abi... 1.2 sec (running source_only_abi[1.0 sec])
   - //third-party/java/dx:dx... 3.3 sec (running javac_jar[3.1 sec])

Reviewed By: dinhviethoa

fbshipit-source-id: 367faa7