Number of build threads to use at coordinator and minions can now be specified as a ratio of total via .buckconfig
Summary:
Besides performing builds, the coordinators and minions need available resources to communicate between themselves in a timely manner. To ensure this is the case, making it possible to limit the number of builder threads via two new config options.

Usage (with default values) is as below:

  [stampede]
    coordinator_build_capacity_ratio=0.75
    minion_build_capacity_ratio=0.9

Test Plan: ci. run some builds.

Reviewed By: ruibm

fbshipit-source-id: 2643ac4