Speed up AbstractCxxSymlinkTreeHeaders appendToRuleKey
Summary:
Instead of stringifying every key and wrapping it with "include()", bracket the
entire structure with two constants. This reduces the number of allocations
which was used to generate the altered string.

Also, this field was being set with an ImmutableSortedMap, so propagate the
type to avoid the need to sort it again.

Reviewed By: ttsugriy

fbshipit-source-id: cd5b01b