Add support for duplicate resource name whitelisting
Summary:
This allows certain resources to have duplicate names, even if they are
of types that would normally be banned.  The intention is to allow apps
to ban duplicate resources by default but allow legacy or special cases
where necessary.

Test Plan: Updated unit test.  Will test with a real app as well.

Reviewed By: asp2insp

fbshipit-source-id: 036c8a9