Fixed running tests on iOS with embedded cells buck-out
Summary:
An intermediate target for xctool-zip is created when running the tests.
When we have several cells, the output path of the target might not be computed correctly.
The output path computed from one cell might differ from the output path computed from an other cell.

This diff will create an intermediate target per cell path buy adding a flavor.
The flavor is computed using the sha1 of the relative path of the cell.
We use the relative path to avoid introducing rule key differences.

Test Plan:
Added a test, make sure that it fails when the change is not applied.
With the change is not applied, the test succeeds.

Unit tests + CI

Reviewed By: ttsugriy

fbshipit-source-id: 4e1ced7