Make CxxLink compute the cell roots correctly
Summary:
CxxLink uses the paths to the cell root to sanitize some output.
Currently it computes it in a rather odd way: it gets the roots of all
of its deps' filesystems. This change makes it actually correctly
compute the roots by passing in a CellPathResolver and deriving from
that.

Test Plan: CI

Reviewed By: andrewjcg

fbshipit-source-id: eb3cecb