Move MacroArg and similar to the macros package
Summary:
This is the beginning of a stack of changes with the goal of changing
the way that macros are expanded. Currently, macros are held in
MacroArg/StringWithMacrosArg/WorkerMacroArg and are expanded multiple
times in different contexts. Once to append to rulekeys, and again later
to expand to strings added to the command line. To support those
expansions, these macro-Args hold references to the entire action graph
and the full cell resolver (and other large things). This makes it
rather impractical to support interesting uses (like serialization).

The goal with this stack will be to change the way that we handle macros
to have them be expanded during action graph creation to Args (that
don't hold such broad references). Args already directly support being
added to rulekeys and support delaying work until being added to command
lines just like macros.

This will allow us to invert the dependency from `args depends on macros`
to `macros depends on args`. That, in turn, is the first step to
changing macros to expand to Args instead of expanding to Strings.

Test Plan: CI, this is just moving things around.

Reviewed By: andrewjcg

fbshipit-source-id: 384e0d0