Make BuildFileParseException unchecked and top-level
Summary:
This diff starts the stack of changes dedicated to making Buck error processing better. In particular, we want to make sure exceptions are properly treated in Main and have similar processing/displaying logic.
I had to make some exceptions unchecked. This is a hard trade-off that comes off the existing design of command pattern in Buck. I tried another approach with checked exceptions which are listed on Command.run() method but this leads to another sort of architectural problems. For example, BuildFileParseException indicates that there was a parsing error during execution of a command. However, some commands (like buck cache) do not have parsing step at all and making exception which may never be thrown there a part of a contract is the same misleading as making it totally unchecked.
The alternative design may be to split command execution into phases - like parse, exec, etc, which one may be implemented or not and throw exceptions of appropriate types. This however is another big refactoring which is not even clear at this point because we do not yet know how incrementa evaluation framework will look like.

Test Plan: tests

Reviewed By: ttsugriy

fbshipit-source-id: 0251462