Cause inplace to extract whls because they are not suppose to be zipimported
Summary:
Whls are not suppose to be installed via adding to path, eggs had that
as a feature. But whls have to be "installed" extracted into the path.

So this extracts the whls into a directory inside the linktree.
We use some touch files to prevent re-extraction.

I tested with "time" and after the first run we don't see any noticable delay

Reviewed By: philipjameson

fbshipit-source-id: b46f524