Symlink shared libraries in directory relative to omnibus library for ghci targets
Summary:
Relying on `LD_LIBRARY_PATH` tricks to load dependencies of shared objects makes the `omnibus` library more difficult to relocate. Any binary resolving the libraries from it needs to have `LD_LIBRARY_PATH` set upon startup with the correct value.
This changes construction of the omnibus to use `-rpath` to a directory relative to the omnibus's location. Packaging the omnibus and the directory of symlinks means we no longer need to rely on `LD_LIBRARY_PATH`

Test Plan: Built some ghci targets in fbcode in `opt` mode. Manually removed the `LD_LIBRARY_PATH` stuff, and saw it work.

Reviewed By: andrewjcg

fbshipit-source-id: ef7d93d