Include module dependency hashes when calculating module hash
Summary:
This change extends the logic responsible for getting module hash by adding plugin dependencies' hashes to the final hash.

The algorithm uses recursive approach which is quite easy to implement using `CacheLoader`. The assumption is that the depth of the dependency graph is not big and it won't result in stack overflow. Dependencies among plugins are checked by `PluginManager` during loading so there is no need to check for that condition.

Test Plan: Local testing.

Reviewed By: ttsugriy

fbshipit-source-id: 6e025ea