Make DistBuildState computation and client-side pre-build phase async.
Summary:
As the title says.
This diff makes SuperConsole recognize the BuildJobState computation as a part of the distributed build, hence getting rid of the "building..." phase in a stampede build without any more information. (Earlier we were posting the DistBuildStarted event *after* computing the file hashes and serializable build job state.)
This change is also required for the upcoming local coordinator mode for stampede.

Test Plan: CI

Reviewed By: alisdair04

fbshipit-source-id: dad5773