Coordinator no longer distributes uncacheable targets to minions
Summary:
This is a redunant step, as uncacheables cannot be shared between minions, nor can they be shared with the client that kicked off the distributed build.

Uncacheable paths from the root target increase the overall distributed build time with no benefit.

Where a cacheable requires an uncacheable, this will still be built on the minion where it is needed anyway.

Test Plan: new unit tests

Reviewed By: shivanker

fbshipit-source-id: bee14d5