`StringWithMacrosArg`: support expansion without consumer
Summary:
Adds functionality to `StringWithMacrosArg` to expand directly to a string, optionally with an additional transform for macro expansions.

`StringWithMacrosArg` is a useful higher level object that is also interesting in contexts where rule key calculation and macro expansion are needed, but not the `Arg` functionality (passing a consumer).

Test Plan: Existing tests for `StringWithMacrosArg` still pass, the same tests are run for `StringWithMacrosAndExpanders`.

Reviewed By: jeanlauliac

fbshipit-source-id: 1372100