classpath() function for deps query traverses android_resource rules
Summary:
We have several instances in the codebase where an `android_library` which provides a `View` implementation is only accessible via an edge from an `android_resource` rule which uses that `View`. This means that the `classpath` function for `deps_query` needs to traverse the `android_resource` rule in order to return a correct closure.
This diff modifies the `GraphEnhancementQueryEnvironment` to traverse `android_resource` rules when evaluating a classpath query.

Test Plan: New integration tests

Reviewed By: dreiss

fbshipit-source-id: 5cad37f