Make gtest return a failure if an exit code >128 is detected
Summary: This makes it so that if gtest dies unexpectedly after running tests, that we record that the test failed. This can often happen with destruction of singletons/objects after "SUCCESS" has been written to screen. This logic is simple, but should catch common cases. If exitCode > 128 (or can't be parsed; it should be able to since it's written in AbstractTestStep) then we say a really bad error happened with the runner so it can be inspected

Test Plan:
Updated unit tests
Verified that a test program would report success with the internal runner before, and failed afterward. Sample project was:

```
#include <gtest/gtest.h>
#include <memory>

class DtorAbort {
public:
  ~DtorAbort() { assert(0 && "I don't think so."); }
};

class DtorFailTest : public ::testing::TestWithParam<std::shared_ptr<DtorAbort>> {
};

TEST_P(DtorFailTest, TestOne) {
  EXPECT_TRUE(true);
  //EXPECT_TRUE(false);
}

INSTANTIATE_TEST_CASE_P(Boom,
                        DtorFailTest,
                        ::testing::Values(std::make_shared<DtorAbort>()));
```

And now returns an error saying signal 6 was caught.

Reviewed By: sbalabanov

fbshipit-source-id: ff2ef9a