Ensure client waits for all build rule finished events before proceeeding
Summary:
At the end of the build the current behavior is to immediately unlock all locally waiting build rules for cache fetches.

This can potentially result in cache misses if the cache hasn't been synchronized yet.

New behavior is that Stampede ensures that at the end of the build we wait for all build rule events to have been received (via a new marker event), and then waits for all received events to be flushed (including waiting for cache synchronization) before proceeding.

Test Plan: ci

Reviewed By: shivanker

fbshipit-source-id: 6d60f22