Add handling for argument file completition
Summary: Argument files are useful ways to collect repeated arguments. Add some bash_completion niceness to make it easier to discover

Test Plan:
Testing it on myself. The version I have locally uses a `find <partial path> -maxdepth 1` to generate
But I learned that apparently there are decent tools for this. A bit of silly handling for the `@` prefix (yes,
I know about the -P arg, but the other handling is easier if I do not need to keep stripping the first character).

Reviewed By: philipjameson

fbshipit-source-id: 6ed4d04