Fix edge case for inner class with generic enclosing instance
Summary:
I previously thought you couldn't get a `DeclaredType` with an enclosing type if there
were any `ErrorType`s involved in the enclosing type. That's not true -- the enclosing
type can be a generic with `ErrorType` type arguments.

See the updated comment and the test changes for details.

Test Plan: Updated unit test

Reviewed By: dreiss

fbshipit-source-id: 976034f