`JsBundle`: Add user-provided json data arg
Summary:
The JS rules communicate with the worker using JSON messages. This adds the ability to specify a custom JSON property for `JsBundle` that is passed through verbatim.

This will allow us to get rid of the `libraryGroups` param, which is tied to the logic of our own worker (not generic).

Test Plan: - new tests

Reviewed By: fromcelticpark

fbshipit-source-id: 8faa999