Enable individual rules to choose whether they're compiled against ABI or full jars
Summary:
In some cases with `source-only-abi`, we'll want rules to be able to compile against `source-abi`s
instead (cases where the `source-only-abi` is correct for most but not all purposes). This is a
step in that direction that can also be useful for cases where any kind of ABI is wrong.

Test Plan: CI

Reviewed By: asp2insp

fbshipit-source-id: 9420e61