Fixed issue where input were not added to the rulekey of Archive when embedded cell buck-out is enabled
Summary:
When embedded buck-out is enabled, output file of rules of a cell are located in [root-buck-out]/cells/[cell-name].
The result of that was that the cell / filesystem the file belongs to was incorrectly calculated.
The result of that is that the file hash was not taken in account when computing the input based rule key.

This diff improves how the cell / filesystem a file belons to is calculated when embedded cell buck-out is enabled.

Test Plan: CI

Reviewed By: ttsugriy

fbshipit-source-id: c9b66d8