Use ToolchainInstantiationException to explain why a toolchain cannot be created
Summary:
The toolchain design allows some toolchains to be not present. At the same time in some cases toolchains need to provide an explanation why they cannot be created. The existing design is not flexible enough to satisfy both situations - toolchain creation either fails with an exception or indicates that toolchain is absent.

This change introduces `ToolchainInstantiationException` which should be used to indicate a reason why a toolchain cannot be created. This exception is thrown when a client is requesting a toolchain unconditionally (by calling `getByName`). This exception is not thrown in `isToolchainPresent` and `getByNameIfPresent`.

`DefaultToolchainProvider` also keeps track of failed toolchains to avoid calling toolchain creation after first failure.

Test Plan: None of the existing toolchains use the new exception so I just verified buck behavior with both valid and invalid Android NDK configuration when building Android and non-Android targets.

Reviewed By: ttsugriy

fbshipit-source-id: 09bab7e